Mowe showMeteoData function to separate file
